refactor: 移動 BombStateEnum 到 types
