+++ scripts/types/bombstate/BombStateEnum.ts
